Fix tests, added back native library resolver for MacOS and linux, other span based api changes.
This should cleanup the api as well as fix running the code on non-Windows systems as well as make tests work cross platform.

As for Windows, I have no idea how to resolve possible issues with loading 32 bit or 64 bit zlib dlls accidentally within the wrong bitness other than bundling the windows versions of the dlls with the library.

Signed-off-by: AraHaan <seandhunt_7@yahoo.com>